+++ src/consts/chains.consts.ts
@@ -1,0 +1,11 @@
+import { ChainInfo } from '@/types'
+
+export const POLYGON_MUMBAI_CHAIN: ChainInfo = {
+  rpcUrl: 'https://matic-mumbai.chainstacklabs.com',
+  blockExplorerUrl: 'https://mumbai.polygonscan.com',
+}
+
+export const Q_TESTNET_CHAIN: ChainInfo = {
+  rpcUrl: 'https://rpc.qtestnet.org',
+  blockExplorerUrl: 'https://explorer.qtestnet.org/',
+}