+++ src/consts/index.ts
@@ -1,2 +1,3 @@
 export * from '@/consts/promocode'
+export * from '@/consts/chains.consts'
 export * from '@/consts/nft'