+++ src/enums/chains.enum.ts
@@ -3,6 +3,17 @@ export enum ETHEREUM_CHAINS {
   ropsten = '3',
   kovan = '42',
   goerli = '5',
+  ethereum = '1',
+}
+
+export enum POLYGON_CHAINS {
+  mumbai = '80001',
+  mainnet = '137',
+}
+
+export enum Q_CHAINS {
+  mainet = '35441',
+  testnet = '35443',
 }
 
 export enum SOLANA_CHAINS {