+++ src/enums/networks.enum.ts
@@ -1,0 +1,6 @@
+export enum NETWORKS {
+  POLYGON = 'polygon',
+  ETHEREUM = 'ethereum',
+  Q = 'q',
+  UNSUPPORTED = 'unsupported',
+}