+++ src/types/chain.types.ts
@@ -1,0 +1,14 @@
+import { POLYGON_CHAINS, ETHEREUM_CHAINS, Q_CHAINS } from '@/enums'
+
+export type NativeCurrency = {
+  name: string
+  symbol: string
+  decimals: number
+}
+
+export type ChainID = POLYGON_CHAINS | ETHEREUM_CHAINS | Q_CHAINS
+
+export type ChainInfo = {
+  rpcUrl: string
+  blockExplorerUrl: string
+}